% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JusticeRegress.R
\name{JusticeRegress}
\alias{JusticeRegress}
\title{JusticeRegress: Fits a Linear Regression model}
\usage{
JusticeRegress(Y, X, alpha = 0.05)
}
\arguments{
\item{Y}{a numeric vector of the responses}

\item{X}{a numeric design matrix for the model.It is a column vector for Simple Linear Regression and matrix for Multiple Linear Regression. See details for more information.}

\item{alpha}{the significance level for your hypotheses tests}
}
\value{
A list including elements
\item{Estimates}{a dataframe for the estimated coefficients, standard errors, t values, and their p values}
\item{Confidence_Intervals}{confidence intervals for the regression coefficients}
\item{Fitted_values}{a vector of the fitted values from the model}
}
\description{
This comprehensive function is used to fit both Simple and Multiple Linear Regression models. The package is designed to be user-friendly.It provides
estimates of the Regression coefficients, standard error, t values, p values, confidence intervals
for the estimates and the fitted values from the model.
}
\details{
The function uses the Ordinary Least Squares (OLS) method to estimate regression coefficients. It internally computes standard errors, t values, and p values for each coefficient. Confidence intervals are calculated based on the specified alpha level. Fitted values are obtained directly from the OLS model.When fitting an SLR model, X should be a column vector otherwise X is a matrix where each column represents a specific variable in the model.
}
\examples{
#Example 1 for Simple Linear Regression
#mtcars dataset in R
data(mtcars)
X=cbind(mtcars$hp) #For Simple Linear Regression X must be a column vector
Y=mtcars$mpg
JusticeRegress(Y, X, alpha = 0.05)

#Example 2 for Multiple Linear Regression
Y=c(27,29,23,20,21)
X=cbind(c(4,7,6,2,3),c(0,1,1,0,0),c(1,1,0,0,1))
JusticeRegress(Y,X) #without specifying alpha.Still uses 0.05

}
\references{
{Chambers, J. M. (1992) Linear models. Chapter 4 of Statistical Models in S eds J. M. Chambers and T. J. Hastie, Wadsworth & Brooks/Cole.}
}
\author{
Justice Akuoko-Frimpong
}
